add apf::Field interface to field reduction
