+++ apf/apfFieldData.h
@@ -11,6 +11,7 @@
 #include <string>
 #include "apfField.h"
 #include "apfShape.h"
+#include "apf.h"  // needed for ReductionOp
 
 namespace apf {
 
@@ -32,46 +33,12 @@ class FieldData
 template <class T>
 class FieldDataOf;
 
-// different reduction operations for Fields
-template <class T>
-class ReductionOp
-{
-  public:
-    virtual T apply(T val1, T val2) const = 0;
-};
-
-template <class T>
-class ReductionSum : public ReductionOp<T>
-{
-  T apply(T val1, T val2) const { return val1 + val2; };
-};
-
-template <class T>
-class ReductionMin : public ReductionOp<T>
-{
-  T apply(T val1, T val2) const { return ( (val1 < val2) ? val1 : val2 ); };
-};
-
-template <class T>
-class ReductionMax : public ReductionOp<T>
-{
-  T apply(T val1, T val2) const { return ( (val1 < val2) ? val2 : val1 ); };
-};
-
-
-/* instantiate (is this necessary with the global consts below?) */
-template class ReductionSum<double>;
-template class ReductionMin<double>;
-template class ReductionMax<double>;
-
-
-
 template <class T>
 void synchronizeFieldData(FieldDataOf<T>* data, Sharing* shr, bool delete_shr=false);
 
 void accumulateFieldData(FieldDataOf<double>* data, Sharing* shr, bool delete_shr=false);
 
-void reduceFieldData(FieldDataOf<double>* data, Sharing* shr, bool delete_shr=false, const apf::ReductionOp<double>& reduce_op=ReductionSum<double>());
+void reduceFieldData(FieldDataOf<double>* data, Sharing* shr, bool delete_shr=false, const ReductionOp<double>& reduce_op=ReductionSum<double>());
 
 template <class T>
 void copyFieldData(FieldDataOf<T>* from, FieldDataOf<T>* to);
